+++ src/env.sh
@@ -27,6 +27,7 @@ _DEFAULT_HEADER_ASCII_ART="false"
 _DEFAULT_SIMPLE_OUTPUT="false"
 _DEFAULT_STOP_ON_FAILURE="false"
 _DEFAULT_SHOW_EXECUTION_TIME="true"
+_DEFAULT_DEV_MODE="false"
 
 : "${BASHUNIT_PARALLEL_RUN:=${PARALLEL_RUN:=$_DEFAULT_PARALLEL_RUN}}"
 : "${BASHUNIT_SHOW_HEADER:=${SHOW_HEADER:=$_DEFAULT_SHOW_HEADER}}"
@@ -34,6 +35,7 @@ _DEFAULT_SHOW_EXECUTION_TIME="true"
 : "${BASHUNIT_SIMPLE_OUTPUT:=${SIMPLE_OUTPUT:=$_DEFAULT_SIMPLE_OUTPUT}}"
 : "${BASHUNIT_STOP_ON_FAILURE:=${STOP_ON_FAILURE:=$_DEFAULT_STOP_ON_FAILURE}}"
 : "${BASHUNIT_SHOW_EXECUTION_TIME:=${SHOW_EXECUTION_TIME:=$_DEFAULT_SHOW_EXECUTION_TIME}}"
+: "${BASHUNIT_DEV_MODE:=${DEV_MODE:=$_DEFAULT_DEV_MODE}}"
 
 function env::is_parallel_run_enabled() {
   [[ "$BASHUNIT_PARALLEL_RUN" == "true" ]]
@@ -59,6 +61,10 @@ function env::is_show_execution_time_enabled() {
   [[ "$BASHUNIT_SHOW_EXECUTION_TIME" == "true" ]]
 }
 
+function env::is_dev_mode_enabled() {
+  [[ "$BASHUNIT_DEV_MODE" == "true" ]]
+}
+
 function env::find_terminal_width() {
   local cols=""
 
@@ -73,6 +79,7 @@ function env::find_terminal_width() {
   echo "${cols:-$_DEFAULT_TERMINAL_WIDTH}"
 }
 
+TEMP_DIR_PARALLEL_TEST_SUITE="/tmp/bashunit/parallel/${_OS:-Unknown}"
 TERMINAL_WIDTH="$(env::find_terminal_width)"
 FAILURES_OUTPUT_PATH=$(mktemp)
 CAT="$(which cat)"