+++ src/main.sh
@@ -15,8 +15,21 @@ function main::exec_tests() {
     exit 1
   fi
 
+  # Trap SIGINT (Ctrl-C) and call the cleanup function
+  trap main::cleanup SIGINT
+
+  if env::is_parallel_run_enabled && check_os::is_alpine; then
+    printf "%sWarning: Parallel test execution on Alpine Linux is currently" "${_COLOR_INCOMPLETE}"
+    printf "in a beta stage.\nThis means there may be unresolved issues, "
+    printf "particularly involving race conditions.%s\n" "${_COLOR_DEFAULT}"
+  fi
+
   console_header::print_version_with_env "$filter" "${test_files[@]}"
   runner::load_test_files "$filter" "${test_files[@]}"
+  if env::is_parallel_run_enabled; then
+    wait
+  fi
+
   console_results::print_failing_tests_and_reset
   console_results::render_result
   exit_code=$?
@@ -30,8 +43,15 @@ function main::exec_tests() {
   fi
 
   cleanup_temp_files
-
   exit $exit_code
+}
+
+function main::cleanup() {
+  printf "%sCaught Ctrl-C, killing all child processes...%s\n"  "${_COLOR_SKIPPED}" "${_COLOR_DEFAULT}"
+  # Kill all child processes of this script
+  pkill -P $$
+  cleanup_temp_files
+  exit 1
 }
 
 function main::exec_assert() {